More int64 changes

+++ tensorflow_quantum/core/ops/math_ops/tfq_inner_product.cc
@@ -227,8 +227,8 @@ class TfqInnerProductOp : public tensorflow::OpKernel {
     const int output_dim_internal_size = output_tensor->dimension(1);
 
     auto DoWork = [&](int64_t start, int64_t end) {
-      int old_batch_index = -2;
-      int cur_batch_index = -1;
+      int64_t old_batch_index = -2;
+      int64_t cur_batch_index = -1;
       uint64_t largest_nq = 1;
       int cur_internal_index;
